+++ docs/changelog.rst
@@ -26,6 +26,7 @@ New
 - Added the option to use landuse/landcover data combined with a reclass table to `SfincsModel.setup_constant_infiltration`.  PR #103
 - Enabled to provide locations only (so no timeseries) for `SfincsModel.setup_waterlevel_forcing` and `SfincsModel.setup_discharge_forcing` PR #104
 - New optional buffer argument in  `SfincsModel.setup_discharge_forcing` to select gauges around boundary only. PR #104
+- New functionality within `SfincsModel.setup_structures` to use high resolution dem for weir elevation
 
 Changed
 -------